refactor(tasks): carve the package into six focused submodules
``orchestration/tasks/__init__.py`` was a 1557-LOC god module.  Split
it along the seams the audit identified — no logic changes, only
relocation:

- meta.py      — on-disk I/O: the dossier/bookkeeping split-write
                 protocol, path helpers, the dirs task state lives in,
                 load/check/mark helpers.  The foundational layer.
- identity.py  — task ID generation, validation, prefix resolution.
- naming.py    — task-name sanitization + random-name generation.
- query.py     — the ``TaskMeta`` read model and the functions that
                 hydrate it from disk + live container state.
- archive.py   — reading the immutable snapshots a deletion leaves.
- lifecycle.py — create / rename / delete (with archive-on-delete) /
                 stop / login / status — the state-mutating and
                 interactive operations.

Dependency DAG (no cycles): meta ← {identity, query, archive},
naming standalone, query ← identity, lifecycle ← everything.

``__init__.py`` is now a pure re-export hub: every name previously
importable from ``terok.lib.orchestration.tasks`` still is, so no
external caller changed.  ``container_name`` / ``CONTAINER_MODES``
keep being re-exported from core.task_state as before.

Tests: six files updated.  Private-symbol test imports
(``_TASK_ID_CROCKFORD_4_5_RE``, ``_generate_unique_id``,
``_default_categories_for_project``, …) now come from the owning
submodule rather than the package root.  Patch targets for symbols
that moved (``tasks_meta_dir`` → ``tasks.meta``, ``get_all_task_states``
/ ``container_git_diff`` → ``tasks.query``, ``os`` / ``shutil`` →
``tasks.lifecycle``, ``_gen_task_id`` → ``tasks.identity``) redirected
to the submodule that actually performs the lookup.

All 2301 unit tests pass; lint, tach, lint-imports, docstrings (97.4%),
REUSE all green.
