+++ src/terok/lib/orchestration/tasks/__init__.py
@@ -3,24 +3,24 @@
 
 """Task metadata, lifecycle, and query operations.
 
-On disk each task is two sibling files keyed on its ID:
+This package is the single import surface for task operations —
+``from terok.lib.orchestration.tasks import …`` resolves every public
+name below.  The implementation is split into focused submodules:
 
-* ``<task_id>_dossier.json`` — the wire-dossier file shield consumers
-  read, in wire shape (``{project, task, name}``).
-* ``<task_id>_meta.yml`` — terok's internal bookkeeping (mode,
-  workspace, web port, hooks_fired, exit_code, lifecycle state).
-
-Format-as-contract: the JSON file's filename says "dossier" and its
-audience is "anyone consuming shield events", so its shape *is* the
-wire dossier — no projection, no key translation.  The YAML file's
-filename says "meta" and its audience is terok-internal; ruamel keeps
-comments and round-trip ergonomics cheap.
-
-The split is reconciled at the I/O boundary: ``read_task_meta``
-returns one merged dict in internal-storage shape (``project_id``,
-``task_id``, …), ``write_task_meta`` splits a single dict back to
-both files atomically.  Pre-self-describing names (``<id>.json`` /
-``<id>.yml``) are migrated to the new layout in place on first read.
+* [`meta`][terok.lib.orchestration.tasks.meta] — the on-disk I/O
+  boundary: the dossier/bookkeeping split-write protocol, path helpers,
+  and the directories task state lives in.
+* [`identity`][terok.lib.orchestration.tasks.identity] — task ID
+  generation, validation, and prefix resolution.
+* [`naming`][terok.lib.orchestration.tasks.naming] — task name
+  sanitization and random-name generation.
+* [`query`][terok.lib.orchestration.tasks.query] — the ``TaskMeta``
+  read model and the functions that hydrate it from disk + live
+  container state.
+* [`lifecycle`][terok.lib.orchestration.tasks.lifecycle] — create,
+  rename, delete (with archive-on-delete), stop, login, and status.
+* [`archive`][terok.lib.orchestration.tasks.archive] — reading the
+  immutable snapshots a deletion leaves behind.
 
 Container runner functions (``task_run_cli``, ``task_run_headless``,
 ``task_restart``) live in the companion ``task_runners`` module.
@@ -29,1529 +29,103 @@ their presentation tables live in ``task_display``.  Log viewing lives
 in ``task_logs``.
 """
 
-import json
-import os
-import re
-import secrets
-import shutil
-import string
-import warnings
-from collections.abc import Iterator
-from dataclasses import dataclass
-from datetime import UTC, datetime
-from pathlib import Path
-from typing import Any
-
-from terok.lib.integrations.executor import AgentRunner
-
-from ...core import runtime as _rt
-from ...core.config import archive_dir
-from ...core.paths import core_state_dir
-from ...core.projects import ProjectConfig, load_project
-from ...core.task_display import STATUS_DISPLAY, mode_info
-from ...core.task_state import (
-    CONTAINER_MODES,
-    TaskState,
-    container_name,
-    effective_status,
+# ``container_name`` / ``CONTAINER_MODES`` originate in core.task_state;
+# tasks has historically re-exported them as part of its public surface.
+from ...core.task_state import CONTAINER_MODES, container_name  # noqa: F401
+from .archive import (  # noqa: F401 — re-exported public API
+    ArchivedTask,
+    list_archived_tasks,
+    task_archive_list,
+    task_archive_logs,
 )
-from ...core.work_status import read_work_status
-from ...util.ansi import (
-    green as _green,
-    red as _red,
-    supports_color as _supports_color,
-    yellow as _yellow,
+from .identity import (  # noqa: F401 — re-exported public API
+    is_task_id,
+    normalize_task_id_input,
+    resolve_task_id,
 )
-from ...util.emoji import render_emoji
-from ...util.fs import archive_timestamp, create_archive_dir, ensure_dir
-from ...util.host_cmd import WORKSPACE_DANGEROUS_DIRNAME
-from ...util.logging_utils import _log_debug
-from ...util.yaml import dump as _yaml_dump, load as _yaml_load
-from ..container_exec import container_git_diff
-
-# ---------- Task meta file format ----------
-#
-# Per-task state lives in two sibling files in ``tasks_meta_dir``,
-# named with a self-describing suffix that signals their audience:
-#
-# * ``<task_id>_dossier.json`` — *the dossier*.  Exactly the keys the
-#   clearance UI renders (``project`` / ``task`` / ``name``).  This
-#   file is the contract terok-shield reads on every event from
-#   inside the container's user namespace using nothing but the
-#   stdlib (no ruamel, no terok package), forwarding its contents
-#   verbatim onto every block / shield-up / shield-down.  JSON,
-#   dict-of-strings, wire shape — the format is part of the
-#   contract.
-#
-# * ``<task_id>_meta.yml`` — terok's internal bookkeeping.  Mode,
-#   workspace, web port, web token, hooks_fired, exit_code,
-#   lifecycle state — everything the orchestrator cares about and
-#   nobody else does.  YAML to keep round-trip comments and human
-#   edits cheap; single consumer, so ruamel on the hot path is fine.
-#
-# Format-as-contract: when a file's name says JSON, its audience is
-# "anyone consuming shield events"; when it says YAML, the audience
-# is terok-internal.  No question which field belongs where — the
-# filename answers.
-
-#: Suffix of the wire-dossier JSON file (full name: ``<task_id>_dossier.json``).
-_DOSSIER_SUFFIX = "_dossier.json"
-
-#: Suffix of the orchestrator-bookkeeping YAML file (full name: ``<task_id>_meta.yml``).
-_META_SUFFIX = "_meta.yml"
-
-# Subset of in-memory meta keys that belong on the wire (and thus in
-# the JSON dossier file).  Stored under internal names — translated
-# at the I/O boundary by ``_DOSSIER_TO_WIRE`` / ``_DOSSIER_FROM_WIRE``.
-_DOSSIER_INTERNAL_KEYS = ("project_id", "task_id", "name")
-_DOSSIER_TO_WIRE = {"project_id": "project", "task_id": "task", "name": "name"}
-_DOSSIER_FROM_WIRE = {v: k for k, v in _DOSSIER_TO_WIRE.items()}
-
-
-def dossier_path(meta_dir: Path, task_id: str) -> Path:
-    """Path to the wire-dossier JSON file — what shield consumers read.
-
-    The OCI ``dossier.meta_path`` annotation points operators at *this*
-    file.  Companion bookkeeping lives at the ``_meta.yml`` sibling.
-    """
-    return meta_dir / f"{task_id}{_DOSSIER_SUFFIX}"
-
-
-def meta_path(meta_dir: Path, task_id: str) -> Path:
-    """Path to the orchestrator-bookkeeping YAML — terok-internal state.
-
-    Holds everything except the wire-dossier triple.  Single consumer
-    (terok itself), so ruamel round-tripping is fine.
-    """
-    return meta_dir / f"{task_id}{_META_SUFFIX}"
-
-
-def _migrate_legacy_filenames(meta_dir: Path, task_id: str) -> None:
-    """Rename pre-self-describing meta files in place — one-shot, idempotent.
-
-    Earlier iterations stored the dossier and meta as ``<task_id>.json``
-    and ``<task_id>.yml``; the new names spell out their audience.
-    Rename eagerly so every other read/write/iter helper deals with the
-    canonical layout exclusively.  Idempotent: a second call is a no-op.
-    """
-    pairs = (
-        (meta_dir / f"{task_id}.json", dossier_path(meta_dir, task_id)),
-        (meta_dir / f"{task_id}.yml", meta_path(meta_dir, task_id)),
-    )
-    for old, new in pairs:
-        if old.is_file() and not new.is_file():
-            old.rename(new)
-
-
-def read_task_meta(meta_dir: Path, task_id: str) -> dict | None:
-    """Compose the orchestrator's logical task-meta dict from the on-disk pair.
-
-    Reads the wire-dossier JSON and the internal YAML, translates the
-    JSON's wire keys back to internal storage names, and returns the
-    union.  Either file may be absent.  Returns ``None`` only when
-    neither file is on disk.
-
-    A pre-self-describing layout (``<task_id>.json`` / ``<task_id>.yml``)
-    is migrated to the new names in place before the read so callers
-    never see the legacy paths.  Likewise a single-JSON layout that
-    mixed wire dossier with bookkeeping (the brief intermediate state
-    on this branch before the split) is detected and re-split here.
-    """
-    _migrate_legacy_filenames(meta_dir, task_id)
-    json_path = dossier_path(meta_dir, task_id)
-    yml_path = meta_path(meta_dir, task_id)
-
-    if not (json_path.is_file() or yml_path.is_file()):
-        return None
-
-    json_data: dict = {}
-    if json_path.is_file():
-        text = json_path.read_text(encoding="utf-8")
-        json_data = json.loads(text) if text.strip() else {}
-
-    # Pre-split single-JSON layout (this branch's earlier iteration)
-    # mixed wire dossier with bookkeeping.  Detect, split, normalise.
-    spillover = {k: v for k, v in json_data.items() if k not in _DOSSIER_FROM_WIRE}
-    if spillover:
-        json_data = {k: v for k, v in json_data.items() if k in _DOSSIER_FROM_WIRE}
-
-    yml_data: dict = {}
-    if yml_path.is_file():
-        yml_data = _to_plain(_yaml_load(yml_path.read_text(encoding="utf-8")) or {})
-
-    # Spillover came from a caller writing the pre-split layout — treat
-    # it as more authoritative than any YAML on disk, since that YAML
-    # was written before the latest mutation and may be stale.
-    if spillover:
-        yml_data = {**yml_data, **spillover}
-
-    merged: dict = dict(yml_data)
-    for wire_key, value in json_data.items():
-        if wire_key in _DOSSIER_FROM_WIRE:
-            merged[_DOSSIER_FROM_WIRE[wire_key]] = value
-
-    # Backfill ``project_id`` from the meta-dir path for legacy records
-    # that predate the field landing in TaskMeta.  Without this, a
-    # task-rename on a legacy task lands a half-populated dossier (no
-    # ``project``) on the wire until some other code path happens to
-    # set it.  Path layout is ``<state>/projects/<project_id>/tasks``.
-    backfill_needed = not merged.get("project_id")
-    if backfill_needed and meta_dir.parent.parent.name == "projects":
-        merged["project_id"] = meta_dir.parent.name
-
-    if spillover or backfill_needed:
-        # Normalise on disk so the next read takes the fast path.
-        write_task_meta(dossier_path(meta_dir, task_id), merged)
-
-    return merged
-
-
-def write_task_meta(dossier_handle: Path, meta: dict) -> None:
-    """Atomic split-write: ``_dossier.json`` + ``_meta.yml`` in one atomic step each.
-
-    *dossier_handle* is a dossier-file handle — what
-    [`dossier_path`][terok.lib.orchestration.tasks.dossier_path] returns
-    and what [`load_task_meta`][terok.lib.orchestration.tasks.load_task_meta]
-    hands back.  Pre-self-describing handles (``<id>.json``) are accepted
-    too and canonicalized on write.  The companion bookkeeping path is
-    derived by swapping the suffix.  Atomic-rename writes (``.tmp`` +
-    ``os.replace``) on each file mean a partial write under EINTR /
-    power loss can leave one stale and the other fresh, but never a
-    half-written file — readers always get a parseable shape.
-    """
-    meta_dir, task_id = _dossier_handle_to_dir_and_id(dossier_handle)
-    meta_dir.mkdir(parents=True, exist_ok=True)
-
-    dossier = {
-        wire_key: str(meta[internal_key])
-        for internal_key, wire_key in _DOSSIER_TO_WIRE.items()
-        if meta.get(internal_key)
-    }
-    state = {k: v for k, v in meta.items() if k not in _DOSSIER_INTERNAL_KEYS}
-
-    _atomic_write(
-        dossier_path(meta_dir, task_id),
-        json.dumps(dossier, indent=2, ensure_ascii=False, default=str) + "\n",
-    )
-    _atomic_write(meta_path(meta_dir, task_id), _yaml_dump(state))
-
-
-def _dossier_handle_to_dir_and_id(path: Path) -> tuple[Path, str]:
-    """Decompose a dossier-file handle into ``(meta_dir, task_id)``.
-
-    Accepts both the canonical ``<task_id>_dossier.json`` and the
-    pre-self-describing ``<task_id>.json`` so callers caching a handle
-    across the migration boundary keep working.  Raises on anything
-    else — silently inferring a task_id from a foreign filename would
-    let bugs pass undetected.
-    """
-    name = path.name
-    if name.endswith(_DOSSIER_SUFFIX):
-        return path.parent, name[: -len(_DOSSIER_SUFFIX)]
-    if name.endswith(".json"):
-        return path.parent, name[: -len(".json")]
-    raise ValueError(f"not a dossier-file handle: {path}")
-
-
-def _atomic_write(path: Path, text: str) -> None:
-    """Stage to ``<path>.tmp`` and rename — POSIX-atomic on the same filesystem."""
-    tmp = path.with_suffix(path.suffix + ".tmp")
-    tmp.write_text(text, encoding="utf-8")
-    os.replace(tmp, path)
-
-
-def _to_plain(obj: Any) -> Any:
-    """Recursively unwrap ruamel ``CommentedMap`` / ``CommentedSeq`` to plain types.
-
-    Round-trip-mode YAML hands back commented containers that are dict/list
-    subclasses.  ``json.dumps`` accepts them but the merged dict should be
-    fully plain so downstream readers don't inadvertently inherit metadata
-    that no longer applies.
-    """
-    if isinstance(obj, dict):
-        return {str(k): _to_plain(v) for k, v in obj.items()}
-    if isinstance(obj, list):
-        return [_to_plain(v) for v in obj]
-    return obj
-
-
-def iter_task_ids(meta_dir: Path) -> Iterator[str]:
-    """Yield every task ID with at least one meta file in *meta_dir*.
-
-    Recognises both the canonical layout (``<id>_dossier.json`` /
-    ``<id>_meta.yml``) and the pre-self-describing layout
-    (``<id>.json`` / ``<id>.yml``) — the latter is migrated lazily on
-    the next ``read_task_meta`` call, so transient mixed states are
-    yielded under their canonical task ID without double-counting.
-    """
-    if not meta_dir.is_dir():
-        return
-    seen: set[str] = set()
-    for path in meta_dir.iterdir():
-        if not path.is_file() or path.name.endswith(".tmp"):
-            continue
-        tid = _task_id_from_filename(path.name)
-        if tid and tid not in seen:
-            seen.add(tid)
-            yield tid
-
-
-def _task_id_from_filename(name: str) -> str:
-    """Strip the meta-file suffix to recover the task ID, ``""`` if unrecognised."""
-    for suffix in (_DOSSIER_SUFFIX, _META_SUFFIX, ".json", ".yml"):
-        if name.endswith(suffix):
-            return name[: -len(suffix)]
-    return ""
-
-
-def task_exists(project_id: str, task_id: str) -> bool:
-    """Return ``True`` if any task-meta file exists for ``(project_id, task_id)``.
-
-    Considers both the canonical (``<task_id>.dossier.json`` /
-    ``<task_id>.meta.yml``) and legacy (``<task_id>.json`` /
-    ``<task_id>.yml``) on-disk layouts.
-    """
-    meta_dir = tasks_meta_dir(project_id)
-    return (
-        dossier_path(meta_dir, task_id).is_file()
-        or meta_path(meta_dir, task_id).is_file()
-        or (meta_dir / f"{task_id}.json").is_file()
-        or (meta_dir / f"{task_id}.yml").is_file()
-    )
-
-
-# ---------- Task IDs ----------
-
-_TASK_ID_HEAD_CHARS = "ghjkmnpqrstvwxyz"
-"""Crockford-legal lowercase letters outside hex (``g-z`` minus ``i l o u``, 16 chars).
-
-First character of every task ID, chosen so the ID is unmistakably non-hex
-within the first character — disambiguating terok task IDs from podman
-container IDs, git SHAs, and other hex blobs at a glance.
-"""
-
-_TASK_ID_BODY_CHARS = "0123456789abcdefghjkmnpqrstvwxyz"
-"""Full Crockford base32 alphabet, lowercase (``0-9 a-z`` minus ``i l o u``, 32 chars)."""
-
-_TASK_ID_LEN = 5
-"""Task ID length.  16 · 10 · 32^3 ≈ 5.2M ids per project (~22.3 bits,
-effectively Crockford-4.5 by entropy) — ample for the retry-on-collision
-loop given realistic project sizes."""
-
-_TASK_ID_CROCKFORD_4_5_RE = re.compile(r"[ghjkmnp-tv-z][0-9][0-9a-hjkmnp-tv-z]{3}")
-"""Canonical full-form validator for current-format task IDs.
-
-Called *Crockford-4.5* because the structural-signature chars at
-positions 1–2 (16-letter head, 10 digits) carry only 4.64 bits
-instead of 10 — total entropy 22.32 bits, equivalent to 4.46 full
-Crockford chars rather than 5.
-"""
-
-_TASK_ID_PREFIX_RE = re.compile(r"[ghjkmnp-tv-z](?:[0-9][0-9a-hjkmnp-tv-z]{0,3})?")
-"""Prefix-match regex for a current-format task ID (1 to `_TASK_ID_LEN` chars)."""
-
-_LEGACY_HEX_TASK_ID_PREFIX_RE = re.compile(r"[0-9a-f]{1,8}")
-"""Prefix-match regex for pre-0.8.0 hex task IDs.  Deprecated in 0.8.0; removal in 0.9.0."""
-
-_LEGACY_HEX_TASK_ID_FULL_RE = re.compile(r"[0-9a-f]{8}")
-"""Full-form validator for pre-0.8.0 hex task IDs.  Deprecated; removal in 0.9.0."""
-
-_TASK_ID_AMBIGUOUS_LETTERS = "ilo"
-"""Crockford's visually ambiguous letters: ``I``/``L`` → ``1`` and ``O`` → ``0``.
-
-See https://www.crockford.com/base32.html.  We encode only in canonical
-lowercase, but accept these substitutions at user-facing entry points to
-widen the input surface.
-"""
-
-_TASK_ID_INPUT_TRANSLATE = str.maketrans(_TASK_ID_AMBIGUOUS_LETTERS, "110")
-"""Translate table matching `_TASK_ID_AMBIGUOUS_LETTERS`."""
-
-
-def normalize_task_id_input(raw: str) -> str:
-    """Collapse user-input variants to the canonical lowercase form.
-
-    Strips hyphens, lowercases, and applies the Crockford
-    ``I/L → 1``, ``O → 0`` substitutions.  The result is still subject
-    to `_TASK_ID_PREFIX_RE` downstream — this only widens what
-    we accept, never what we emit.
-
-    **Call-site discipline:** only call this at user-interactive CLI
-    boundaries — argparse dispatch handlers and argcomplete completers.
-    Internal code paths (``lib/*``, ``tui/*``, TUI pickers, clearance,
-    anything reading task IDs from disk, OCI annotations, or runtime
-    state) always work with canonical lowercase IDs and must *never*
-    re-normalise.  Leaking this tolerance inward quietly defeats the
-    "we encode only in canonical form" invariant.
-    """
-    return raw.replace("-", "").lower().translate(_TASK_ID_INPUT_TRANSLATE)
-
-
-def is_task_id(text: str) -> bool:
-    """Return True if *text* is a well-formed task ID (current or legacy)."""
-    return bool(
-        _TASK_ID_CROCKFORD_4_5_RE.fullmatch(text) or _LEGACY_HEX_TASK_ID_FULL_RE.fullmatch(text)
-    )
-
-
-def _gen_task_id() -> str:
-    """Return a fresh task ID: Crockford-letter head + digit + 3 Crockford chars."""
-    alphabets = (_TASK_ID_HEAD_CHARS, string.digits) + (_TASK_ID_BODY_CHARS,) * (_TASK_ID_LEN - 2)
-    tid = "".join(map(secrets.choice, alphabets))
-    assert _TASK_ID_CROCKFORD_4_5_RE.fullmatch(tid), tid  # generator-output invariant
-    return tid
-
-
-def _generate_unique_id(existing: set[str]) -> str:
-    """Generate a unique task ID not present in *existing*."""
-    for _ in range(100):
-        candidate = _gen_task_id()
-        if candidate not in existing:
-            return candidate
-    raise RuntimeError("Failed to generate unique task ID after 100 attempts")
-
-
-def _is_initialized(meta: dict) -> bool:
-    """Return True if the task has completed first-boot initialisation."""
-    return "ready_at" in meta
-
-
-def _validate_task_id_prefix(prefix: str) -> None:
-    """Validate *prefix* and warn on the deprecated legacy-hex format.
-
-    Raises ``SystemExit`` if *prefix* matches neither the current
-    Crockford-4.5 format nor the legacy pre-0.8.0 hex format.  Dispatch
-    is unambiguous: Crockford IDs always start with ``[g-z]`` minus
-    ``i l o u``, which is disjoint from hex.
-    """
-    if _TASK_ID_PREFIX_RE.fullmatch(prefix):
-        return
-    if _LEGACY_HEX_TASK_ID_PREFIX_RE.fullmatch(prefix):
-        warnings.warn(
-            f"Task ID {prefix!r} uses the pre-0.8.0 hex format; "
-            "support will be removed in 0.9.0 — recreate the task to "
-            "adopt the current format.",
-            DeprecationWarning,
-            stacklevel=3,
-        )
-        return
-    raise SystemExit(
-        f"Invalid task ID prefix {prefix!r}; "
-        f"expected 1-{_TASK_ID_LEN} Crockford chars (e.g. 'k3v8h')"
-    )
-
-
-def resolve_task_id(project_id: str, prefix: str) -> str:
-    """Resolve a (possibly partial) task ID to its full form.
-
-    The *prefix* is first run through [`normalize_task_id_input`][terok.lib.orchestration.tasks.normalize_task_id_input],
-    so callers may pass uppercase, hyphenated, or ambiguous-letter
-    variants (``K3-V8H``, ``k3v8I``) — they collapse to the canonical
-    lowercase form before validation and lookup.
-
-    Raises ``SystemExit`` with an actionable message on zero or multiple matches.
-    """
-    # Head position is letter-only, so an ``I/L/O`` there can't be a
-    # Crockford body substitution — surface that as a specific error
-    # instead of normalising it into a digit and letting downstream
-    # mistake it for a legacy-hex prefix.
-    stripped = prefix.replace("-", "").lower()
-    if stripped[:1] in _TASK_ID_AMBIGUOUS_LETTERS:
-        raise SystemExit(
-            f"Invalid task ID prefix {prefix!r}; "
-            f"task IDs never start with I, L, or O — "
-            f"expected one of {_TASK_ID_HEAD_CHARS!r}"
-        )
-    prefix = stripped.translate(_TASK_ID_INPUT_TRANSLATE)
-    _validate_task_id_prefix(prefix)
-    meta_dir = tasks_meta_dir(project_id)
-    if not meta_dir.is_dir():
-        raise SystemExit(f"No tasks found for project {project_id}")
-    if task_exists(project_id, prefix):
-        return prefix
-    matches = [tid for tid in iter_task_ids(meta_dir) if is_task_id(tid) and tid.startswith(prefix)]
-    if len(matches) == 1:
-        return matches[0]
-    if not matches:
-        raise SystemExit(f"No task matching '{prefix}' in project {project_id}")
-    raise SystemExit(f"Ambiguous task ID '{prefix}' — matches: {', '.join(sorted(matches))}")
-
-
-def get_task_container_state(project_id: str, task_id: str, mode: str | None) -> str | None:
-    """Get actual container state for a task (TUI helper)."""
-    if not mode:
-        return None
-    cname = container_name(project_id, mode, task_id)
-    return _rt.get_runtime().container(cname).state
-
-
-@dataclass(kw_only=True)
-class TaskMeta(TaskState):
-    """Lightweight metadata snapshot for a single task.
-
-    Inherits lifecycle fields (``container_state``, ``exit_code``,
-    ``deleting``, ``initialized``) from [`TaskState`][terok.lib.core.task_state.TaskState].
-    """
-
-    task_id: str
-    project_id: str = ""
-    """Project the task belongs to.
-
-    Carried in the meta JSON so consumers that don't already know the
-    project (e.g. terok-shield's host-side dossier resolver, which only
-    has the meta-path pointer) can render full ``project/task`` identity
-    without re-deriving it from the on-disk path.  Empty string for
-    pre-this-release meta files; the lifecycle backfills it on the
-    next mutation.
-    """
-    mode: str | None
-    workspace: str
-    web_port: int | None
-    web_token: str | None = None
-    backend: str | None = None
-    preset: str | None = None
-    name: str = ""
-    provider: str | None = None
-    unrestricted: bool | None = None
-    work_status: str | None = None
-    work_message: str | None = None
-    shield_state: str | None = None
-    created_at: str | None = None
-
-    @property
-    def status(self) -> str:
-        """Compute effective status from live container state + metadata."""
-        return effective_status(self)
-
-
-TASK_NAME_MAX_LEN = 60
-"""Maximum length of a sanitized task name."""
-
-
-def sanitize_task_name(raw: str | None) -> str | None:
-    """Sanitize a raw task name into a slug-style identifier.
-
-    Strips whitespace, lowercases, replaces spaces with hyphens,
-    removes characters outside ``[a-z0-9_-]``, collapses consecutive
-    hyphens, strips trailing hyphens, and truncates to
-    ``TASK_NAME_MAX_LEN``.  Returns ``None`` if the result is empty.
-
-    Leading hyphens are preserved so callers can detect and reject them
-    (a name starting with ``-`` looks like a CLI flag).
-    """
-    if raw is None:
-        return None
-    name = raw.strip().lower()
-    name = name.replace(" ", "-")
-    name = re.sub(r"[^a-z0-9_-]", "", name)
-    name = re.sub(r"-{2,}", "-", name)
-    name = name.rstrip("-")
-    name = name[:TASK_NAME_MAX_LEN]
-    return name or None
-
-
-def validate_task_name(sanitized: str) -> str | None:
-    """Return an error message if *sanitized* is not a valid task name, else ``None``.
-
-    A name is invalid if it starts with a hyphen (looks like a CLI flag).
-    Callers should first check for ``None`` from [`sanitize_task_name`][terok.lib.orchestration.tasks.sanitize_task_name]
-    (which indicates the name was empty after sanitization).
-    """
-    if sanitized.startswith("-"):
-        return "name must not start with a hyphen"
-    return None
-
-
-def generate_task_name(project_id: str | None = None) -> str:
-    """Generate a random human-readable task name (e.g. ``talented-toucan``).
-
-    When *project_id* is given, name categories are resolved from config:
-    project ``tasks.name_categories`` → global ``tasks.name_categories``
-    → deterministic 3-category selection based on project ID hash.
-    """
-    import namer
-
-    categories = _resolve_name_categories(project_id) if project_id else None
-    return namer.generate(separator="-", category=categories)
-
-
-def _resolve_name_categories(project_id: str) -> list[str] | None:
-    """Resolve task-name categories: project config → global config → hash default."""
-    from ...core.config import get_task_name_categories
-
-    # 1. Per-project override
-    try:
-        project = load_project(project_id)
-        if project.task_name_categories:
-            return project.task_name_categories
-    except SystemExit:
-        pass
-
-    # 2. Global config
-    global_cats = get_task_name_categories()
-    if global_cats:
-        return global_cats
-
-    # 3. Hash-based default: pick 3 categories deterministically from project ID
-    return _default_categories_for_project(project_id)
-
-
-def _default_categories_for_project(project_id: str) -> list[str]:
-    """Pick 3 categories deterministically based on a hash of the project ID."""
-    import hashlib
-    import random
-
-    import namer
-
-    categories = sorted(namer.list_categories())
-    seed = int(hashlib.md5(project_id.encode(), usedforsecurity=False).hexdigest(), 16)
-    rng = random.Random(seed)
-    return rng.sample(categories, min(3, len(categories)))
-
-
-def get_task_meta(project_id: str, task_id: str) -> TaskMeta:
-    """Return metadata for a single task with live container state.
-
-    Hydrates ``container_state`` from the running container so that
-    ``TaskMeta.status`` reflects current reality rather than stale YAML.
-    Raises ``SystemExit`` if the task metadata file is not found.
-    """
-    meta_dir = tasks_meta_dir(project_id)
-    raw = read_task_meta(meta_dir, task_id)
-    if raw is None:
-        raise SystemExit(f"Unknown task {task_id}")
-    mode = raw.get("mode")
-    tid = str(raw.get("task_id", ""))
-    # Hydrate live container state only for tasks that have actually been started
-    live_state: str | None = None
-    if mode is not None:
-        try:
-            cname = container_name(project_id, mode, task_id)
-            live_state = _rt.get_runtime().container(cname).state
-        except Exception:
-            pass
-    # Hydrate work status from agent-config (same logic as _get_tasks)
-    ws_status: str | None = None
-    ws_message: str | None = None
-    if tid:
-        project = load_project(project_id)
-        try:
-            agent_cfg = project.tasks_root / tid / "agent-config"
-            ws = read_work_status(agent_cfg)
-            ws_status = ws.status
-            ws_message = ws.message
-        except Exception:  # noqa: BLE001 — best-effort; agent-config may not exist yet
-            pass
-    return TaskMeta(
-        task_id=tid,
-        # ``or`` (not the dict default) so a migrated record carrying an
-        # empty string still falls back to the path-derived project_id.
-        project_id=raw.get("project_id") or project_id,
-        mode=mode,
-        workspace=raw.get("workspace", ""),
-        web_port=raw.get("web_port"),
-        web_token=raw.get("web_token"),
-        backend=raw.get("backend"),
-        container_state=live_state,
-        exit_code=raw.get("exit_code"),
-        deleting=bool(raw.get("deleting")),
-        initialized=_is_initialized(raw),
-        preset=raw.get("preset"),
-        name=raw["name"],
-        provider=raw.get("provider"),
-        unrestricted=raw.get("unrestricted"),
-        work_status=ws_status,
-        work_message=ws_message,
-        created_at=raw.get("created_at"),
-    )
-
-
-def get_workspace_git_diff(project_id: str, task_id: str, against: str = "HEAD") -> str | None:
-    """Get git diff from a task's workspace via container exec.
-
-    Runs ``git diff`` **inside** the task container rather than on the host,
-    so that even poisoned git hooks only execute within the container sandbox.
-
-    Args:
-        project_id: The project ID
-        task_id: The task ID
-        against: What to diff against (``"HEAD"`` or ``"PREV"``)
-
-    Returns:
-        The git diff output as a string, or ``None`` if failed
-    """
-    try:
-        load_project(project_id)  # validate project exists
-        meta_dir = tasks_meta_dir(project_id)
-        meta = read_task_meta(meta_dir, task_id)
-        if meta is None:
-            return None
-        mode = meta.get("mode")
-        if not mode:
-            return None
-
-        if against == "PREV":
-            return container_git_diff(project_id, task_id, mode, "HEAD~1", "HEAD")
-        return container_git_diff(project_id, task_id, mode, "HEAD")
-
-    except (Exception, SystemExit):
-        return None
-
-
-# ---------- Tasks ----------
-
-
-def tasks_meta_dir(project_id: str) -> Path:
-    """Return the directory containing task metadata files for *project_id*."""
-    return core_state_dir() / "projects" / project_id / "tasks"
-
-
-CONTAINER_TEROK_CONFIG = "/home/dev/.terok"
-"""In-container mount point for the per-task agent-config dir."""
-
-
-def agent_config_dir(project_id: str, task_id: str) -> Path:
-    """Host path of the agent-config dir bind-mounted at `CONTAINER_TEROK_CONFIG`."""
-    return load_project(project_id).tasks_root / str(task_id) / "agent-config"
-
-
-def tasks_archive_dir(project_id: str) -> Path:
-    """Return the directory containing archived task data for *project_id*.
-
-    Lives under the namespace archive tree (``archive/<pid>/tasks/``) so
-    operators can find all archived data in one location.  On project
-    deletion the entire ``archive/<pid>/`` subtree is bundled into the
-    project snapshot and removed.
-    """
-    return archive_dir() / project_id / "tasks"
-
-
-def update_task_exit_code(project_id: str, task_id: str, exit_code: int | None) -> None:
-    """Update task metadata with exit code and final status.
-
-    Args:
-        project_id: The project ID
-        task_id: The task ID
-        exit_code: The exit code from the task, or None if unknown/failed
-    """
-    meta_dir = tasks_meta_dir(project_id)
-    meta = read_task_meta(meta_dir, task_id)
-    if meta is None:
-        return
-    meta["exit_code"] = exit_code
-    write_task_meta(dossier_path(meta_dir, task_id), meta)
-
-
-def _write_task_readme(task_dir: Path) -> None:
-    """Write a README.md explaining the task directory layout and security."""
-    readme = task_dir / "README.md"
-    readme.write_text(
-        "# Task Directory\n"
-        "\n"
-        "## workspace-dangerous/\n"
-        "\n"
-        "This directory contains a git repository checked out from the project\n"
-        "source. It is mounted into the task container at `/workspace`.\n"
-        "\n"
-        "**WARNING: Do not execute code or run git commands in this directory\n"
-        "from the host.** The container has full write access and could have\n"
-        "rewritten git hooks, checked in malicious scripts, or otherwise\n"
-        "poisoned the repository contents.\n"
-        "\n"
-        "The safe way to interact with the repository is through the **git\n"
-        "gate** — a separate, host-controlled bare repo that agents push to.\n"
-        "Even online-mode agents can be instructed to mirror their work to\n"
-        "the gate.\n",
-        encoding="utf-8",
-    )
-
-
-def _task_new(project: ProjectConfig, *, name: str | None = None) -> str:
-    """Create a new task with a fresh workspace.  Returns the task ID."""
-    if name is not None:
-        task_name = sanitize_task_name(name)
-        if task_name is None:
-            raise SystemExit(f"Invalid task name: {name!r}")
-        err = validate_task_name(task_name)
-        if err:
-            raise SystemExit(f"Invalid task name: {err}")
-    else:
-        task_name = generate_task_name(project.id)
-    tasks_root = project.tasks_root
-    ensure_dir(tasks_root)
-    meta_dir = tasks_meta_dir(project.id)
-    ensure_dir(meta_dir)
-
-    existing = set(iter_task_ids(meta_dir))
-    next_id = _generate_unique_id(existing)
-
-    ws = tasks_root / next_id
-    ensure_dir(ws)
-
-    workspace_dir = ws / WORKSPACE_DANGEROUS_DIRNAME
-    ensure_dir(workspace_dir)
-    workspace_dir.chmod(0o700)
-    marker_path = workspace_dir / ".new-task-marker"
-    marker_path.write_text(
-        "# This marker signals that the workspace should be reset to the latest remote HEAD.\n"
-        "# It is created by 'terok task new' and removed by init-ssh-and-repo.sh"
-        " after reset.\n"
-        "# If you see this file in an initialized workspace, something went wrong.\n",
-        encoding="utf-8",
-    )
-
-    _write_task_readme(ws)
-
-    meta = {
-        "project_id": project.id,
-        "task_id": next_id,
-        "name": task_name,
-        "mode": None,
-        "workspace": str(ws),
-        "web_port": None,
-        "created_at": datetime.now(tz=UTC).isoformat(),
-    }
-    write_task_meta(dossier_path(meta_dir, next_id), meta)
-    print(f"Created task {next_id} ({task_name}) in {ws}")
-    return next_id
-
-
-def task_new(project_id: str, *, name: str | None = None) -> str:
-    """Create a new task with a fresh workspace for a project.
-
-    Args:
-        project_id: The project to create the task under.
-        name: Optional human-readable name.  Allowed characters are
-            lowercase letters, digits, hyphens, and underscores.
-            If ``None``, a random slug-style name is generated via
-            [`generate_task_name`][terok.lib.orchestration.tasks.generate_task_name].
-
-    Workspace Initialization Protocol:
-    ----------------------------------
-    Each task gets its own workspace directory that persists across container
-    runs. When a container starts, the init script (init-ssh-and-repo.sh) needs
-    to know whether this is:
-
-    1. A NEW task that should be reset to the latest remote HEAD
-    2. A RESTARTED task where local changes should be preserved
-
-    We use a marker file (.new-task-marker) to signal intent:
-
-    - task_new() creates the marker in the workspace directory
-    - init-ssh-and-repo.sh checks for the marker:
-      - If marker exists: reset to origin/HEAD, then delete marker
-      - If no marker: fetch only, preserve local state
-    - Subsequent container runs on the same task won't see the marker,
-      so local work is preserved
-
-    This handles edge cases like:
-    - Stale workspace from incompletely deleted previous task with same ID
-    - Ensuring new tasks always start with latest code
-    """
-    return _task_new(load_project(project_id), name=name)
-
-
-def _task_rename(project: ProjectConfig, task_id: str, new_name: str) -> None:
-    """Rename a task by updating its metadata file."""
-    meta_dir = tasks_meta_dir(project.id)
-    meta = read_task_meta(meta_dir, task_id)
-    if meta is None:
-        raise SystemExit(f"Unknown task {task_id}")
-    sanitized = sanitize_task_name(new_name)
-    if sanitized is None:
-        raise SystemExit(f"Invalid task name: {new_name!r}")
-    err = validate_task_name(sanitized)
-    if err:
-        raise SystemExit(f"Invalid task name: {err}")
-    meta["name"] = sanitized
-    write_task_meta(dossier_path(meta_dir, task_id), meta)
-    print(f"Renamed task {task_id} to {sanitized}")
-
-
-def task_rename(project_id: str, task_id: str, new_name: str) -> None:
-    """Rename a task by updating its metadata file.
-
-    Sanitizes *new_name* and writes the result to the task's metadata file.
-    Raises ``SystemExit`` if the task is unknown or the sanitized name is invalid.
-    """
-    _task_rename(load_project(project_id), task_id, new_name)
-
-
-def _get_tasks(project_id: str, reverse: bool = False) -> list[TaskMeta]:
-    """Return all task metadata for *project_id*, sorted by task ID."""
-    meta_dir = tasks_meta_dir(project_id)
-    tasks: list[TaskMeta] = []
-    if not meta_dir.is_dir():
-        return tasks
-    try:
-        project = load_project(project_id)
-        tasks_root = project.tasks_root
-    except SystemExit:
-        tasks_root = None
-    for tid_stem in iter_task_ids(meta_dir):
-        if not is_task_id(tid_stem):
-            continue
-        try:
-            meta = read_task_meta(meta_dir, tid_stem)
-            if meta is None:
-                continue
-            tid = str(meta.get("task_id", ""))
-            ws_status = None
-            ws_message = None
-            if tasks_root and tid:
-                agent_cfg = tasks_root / tid / "agent-config"
-                ws = read_work_status(agent_cfg)
-                ws_status = ws.status
-                ws_message = ws.message
-            mode = meta.get("mode")
-            tasks.append(
-                TaskMeta(
-                    task_id=tid,
-                    mode=mode,
-                    workspace=meta.get("workspace", ""),
-                    web_port=meta.get("web_port"),
-                    web_token=meta.get("web_token"),
-                    backend=meta.get("backend"),
-                    exit_code=meta.get("exit_code"),
-                    deleting=bool(meta.get("deleting")),
-                    initialized=_is_initialized(meta),
-                    preset=meta.get("preset"),
-                    name=meta["name"],
-                    provider=meta.get("provider"),
-                    unrestricted=meta.get("unrestricted"),
-                    work_status=ws_status,
-                    work_message=ws_message,
-                    created_at=meta.get("created_at"),
-                )
-            )
-        except Exception as exc:
-            from ...util.logging_utils import log_warning
-
-            log_warning(f"Skipping malformed task metadata file for {tid_stem}: {exc}")
-            continue
-
-    tasks.sort(key=lambda t: t.task_id or "", reverse=reverse)
-    return tasks
-
-
-def get_tasks(project_id: str, reverse: bool = False) -> list[TaskMeta]:
-    """Return all task metadata for *project_id*, sorted by task ID."""
-    return _get_tasks(project_id, reverse=reverse)
-
-
-def get_all_task_states(
-    project_id: str,
-    tasks: list[TaskMeta],
-) -> dict[str, str | None]:
-    """Map each task to its live container state via a single batch query.
-
-    Args:
-        project_id: The project whose containers to query.
-        tasks: List of ``TaskMeta`` instances (must have ``task_id`` and ``mode``).
-
-    Returns:
-        ``{task_id: container_state_or_None}`` dict.
-    """
-    # container_states isn't on the ContainerRuntime Protocol (only the
-    # concrete podman impl) — upstream surface gap; the call works at runtime.
-    container_states = _rt.get_runtime().container_states(project_id)  # type: ignore[attr-defined]
-    result: dict[str, str | None] = {}
-    for t in tasks:
-        if t.mode:
-            cname = container_name(project_id, t.mode, str(t.task_id))
-            result[str(t.task_id)] = container_states.get(cname)
-        else:
-            result[str(t.task_id)] = None
-    return result
-
-
-def task_list(
-    project_id: str,
-    *,
-    status: str | None = None,
-    mode: str | None = None,
-    agent: str | None = None,
-) -> None:
-    """List tasks for a project, optionally filtered by status, mode, or agent preset.
-
-    Status is computed live from podman container state + task metadata.
-    """
-    tasks = get_tasks(project_id)
-
-    # Pre-filter by mode/agent before the podman query to reduce work
-    if mode:
-        tasks = [t for t in tasks if t.mode == mode]
-    if agent:
-        tasks = [t for t in tasks if t.preset == agent]
-
-    if not tasks:
-        print("No tasks found")
-        return
-
-    # Batch-query podman for all container states in one call
-    live_states = get_all_task_states(project_id, tasks)
-    for t in tasks:
-        t.container_state = live_states.get(t.task_id)
-
-    # Filter by effective status (computed live)
-    if status:
-        tasks = [t for t in tasks if effective_status(t) == status]
-
-    if not tasks:
-        print("No tasks found")
-        return
-
-    for t in tasks:
-        t_status = effective_status(t)
-        extra = []
-        if t.mode:
-            extra.append(f"mode={t.mode}")
-        if t.web_port:
-            extra.append(f"port={t.web_port}")
-        if t.work_status:
-            extra.append(f"work={t.work_status}")
-        extra_s = f" [{'; '.join(extra)}]" if extra else ""
-        print(f"- {t.task_id}: {t.name} {t_status}{extra_s}")
-
-
-def _check_mode(meta: dict, expected: str) -> None:
-    """Raise SystemExit if the task's mode conflicts with *expected*."""
-    mode = meta.get("mode")
-    if mode and mode != expected:
-        raise SystemExit(f"Task already ran in mode '{mode}', cannot run in '{expected}'")
-
-
-def load_task_meta(
-    project_id: str, task_id: str, expected_mode: str | None = None
-) -> tuple[dict, Path]:
-    """Load task metadata and optionally validate mode.
-
-    Returns the merged in-memory dict (dossier + bookkeeping in their
-    internal storage shape).  Raises SystemExit if the task is unknown
-    or its mode conflicts with *expected_mode*.
-
-    Returns ``(meta, dossier_path)``: the merged in-memory dict (in
-    its internal storage shape — ``project_id`` / ``task_id`` / …)
-    plus the canonical dossier-file handle for write-back via
-    ``write_task_meta(dossier_path, meta)``.  Raises SystemExit if
-    the task is unknown or its mode conflicts with *expected_mode*.
-    """
-    meta_dir = tasks_meta_dir(project_id)
-    meta = read_task_meta(meta_dir, task_id)
-    if meta is None:
-        raise SystemExit(f"Unknown task {task_id}")
-    if expected_mode is not None:
-        _check_mode(meta, expected_mode)
-    return meta, dossier_path(meta_dir, task_id)
-
-
-def mark_task_deleting(project_id: str, task_id: str) -> None:
-    """Persist ``deleting: true`` to the task's metadata file."""
-    try:
-        meta_dir = tasks_meta_dir(project_id)
-        meta = read_task_meta(meta_dir, task_id)
-        if meta is None:
-            return
-        meta["deleting"] = True
-        write_task_meta(dossier_path(meta_dir, task_id), meta)
-    except Exception as e:
-        _log_debug(f"mark_task_deleting: failed project_id={project_id} task_id={task_id}: {e}")
-
-
-def capture_task_logs(project: ProjectConfig | str, task_id: str, mode: str) -> Path | None:
-    """Capture container logs to the task's ``logs/`` directory on the host.
-
-    Writes stdout/stderr from ``podman logs`` to
-    ``<tasks_root>/<task_id>/logs/container.log``.  Returns the log file
-    path on success, or ``None`` if the container doesn't exist or podman
-    fails.
-
-    *project* may be a [`ProjectConfig`][terok.cli.commands.sickbay.ProjectConfig] or a project-ID string
-    (the string form loads the config internally for backward compat).
-    """
-    if isinstance(project, str):
-        project = load_project(project)
-    task_dir = project.tasks_root / str(task_id)
-    logs_dir = task_dir / "logs"
-    ensure_dir(logs_dir)
-    log_file = logs_dir / "container.log"
-
-    cname = container_name(project.id, mode, task_id)
-    if AgentRunner().capture_logs(cname, log_file, timestamps=True, timeout=60.0):
-        return log_file
-    return None
-
-
-def _archive_task(project: ProjectConfig, task_id: str, meta: dict) -> Path | None:
-    """Archive task metadata and logs before deletion.
-
-    Creates an entry at ``archive/<project_id>/tasks/<ts>_<task_id>[_<name>]/``
-    containing the task metadata YAML and any captured logs.  The archival
-    timestamp is the primary identifier because task numbers and names can
-    be reused after deletion.
-
-    Returns the archive directory path, or ``None`` if archiving failed.
-    """
-    try:
-        task_name = meta.get("name", "")
-        ts = archive_timestamp()
-        # Build archive dir name: timestamp_taskid_name (name may be empty)
-        dir_name = f"{ts}_{task_id}"
-        if task_name:
-            dir_name = f"{dir_name}_{task_name}"
-
-        archive_root = tasks_archive_dir(project.id)
-        archive_dir = create_archive_dir(archive_root, dir_name)
-
-        # Save the *merged* meta as a single self-contained JSON snapshot.
-        # The live-task layout splits dossier (JSON) from state (YAML)
-        # because each file has a different audience; an archive entry
-        # has only one audience (the operator looking at history) and is
-        # never re-read by shield, so the split brings no value here.
-        _atomic_write(
-            archive_dir / "task.json",
-            json.dumps(_to_plain(meta), indent=2, ensure_ascii=False, default=str) + "\n",
-        )
-
-        # Copy logs if they exist
-        task_dir = project.tasks_root / str(task_id)
-        logs_dir = task_dir / "logs"
-        if logs_dir.is_dir():
-            archive_logs_dir = archive_dir / "logs"
-            shutil.copytree(logs_dir, archive_logs_dir, dirs_exist_ok=True)
-
-        _log_debug(f"_archive_task: archived task {task_id} to {archive_dir}")
-        return archive_dir
-    except Exception as e:
-        _log_debug(f"_archive_task: failed to archive task {task_id}: {e}")
-        return None
-
-
-@dataclass
-class TaskDeleteResult:
-    """Outcome of a task deletion — always completes, collects warnings.
-
-    The task is considered deleted regardless (metadata and workspace
-    removed), but individual cleanup steps may fail.  ``warnings``
-    carries human-readable descriptions of any steps that did not
-    complete cleanly.
-    """
-
-    task_id: str
-    warnings: list[str]
-
-
-def _task_delete(project: ProjectConfig, task_id: str) -> TaskDeleteResult:
-    """Delete a task's workspace, metadata, and associated containers."""
-    _log_debug(f"task_delete: start project_id={project.id} task_id={task_id}")
-    warnings: list[str] = []
-
-    workspace = project.tasks_root / str(task_id)
-    meta_dir = tasks_meta_dir(project.id)
-    dossier_file = dossier_path(meta_dir, task_id)
-    meta_file = meta_path(meta_dir, task_id)
-    # Pre-self-describing layout (`<id>.json` / `<id>.yml`) — clean too,
-    # in case the task was created before the rename and never since
-    # touched (no read = no migration).
-    legacy_json = meta_dir / f"{task_id}.json"
-    legacy_yml = meta_dir / f"{task_id}.yml"
-    _log_debug(
-        f"task_delete: workspace={workspace} dossier_file={dossier_file} meta_file={meta_file}"
-    )
-
-    meta = read_task_meta(meta_dir, task_id) or {}
-
-    mode = meta.get("mode")
-    if mode:
-        _log_debug("task_delete: capturing container logs")
-        capture_task_logs(project, task_id, mode)
-
-    if meta:
-        _log_debug("task_delete: archiving task")
-        _archive_task(project, task_id, meta)
-
-    _log_debug("task_delete: revoking gate tokens")
-    from terok.lib.integrations.sandbox import revoke_token_for_task
-
-    try:
-        revoke_token_for_task(project.id, task_id)
-    except Exception as exc:
-        _log_debug(f"task_delete: token revoke failed: {exc}")
-        warnings.append(f"Token revoke failed: {exc}")
-
-    _log_debug("task_delete: calling stop_task_containers")
-    names = [container_name(project.id, mode, str(task_id)) for mode in CONTAINER_MODES]
-    containers_removed = True
-    try:
-        rm_results = _rt.get_runtime().force_remove([_rt.get_runtime().container(n) for n in names])
-    except Exception as exc:
-        _log_debug(f"task_delete: stop_task_containers raised: {exc}")
-        warnings.append(f"Container removal failed: {exc}")
-        rm_results = []
-        containers_removed = False
-    for r in rm_results:
-        if not r.removed:
-            _log_debug(f"task_delete: container {r.name} not removed: {r.error}")
-            warnings.append(f"Container {r.name}: {r.error}")
-            containers_removed = False
-    _log_debug("task_delete: stop_task_containers returned")
-
-    if mode:
-        from ..hooks import run_hook
-
-        run_hook(
-            "post_stop",
-            project.hook_post_stop,
-            project_id=project.id,
-            task_id=task_id,
-            mode=mode,
-            cname=container_name(project.id, mode, task_id),
-            task_dir=workspace,
-            meta_path=meta_file,
-        )
-
-    if workspace.is_dir():
-        _log_debug("task_delete: removing workspace directory")
-        try:
-            shutil.rmtree(workspace)
-            _log_debug("task_delete: workspace directory removed")
-        except Exception as exc:
-            _log_debug(f"task_delete: workspace removal failed: {exc}")
-            warnings.append(f"Workspace removal failed: {exc}")
-
-    # Drop both halves of the on-disk pair plus any legacy single-file
-    # remnants.  Each ``unlink(missing_ok=True)`` is independent so a
-    # half-installed bundle still cleans up everything that was there.
-    paths_to_remove = (dossier_file, meta_file, legacy_json, legacy_yml)
-    if any(p.is_file() for p in paths_to_remove):
-        _log_debug("task_delete: removing metadata files")
-        try:
-            for p in paths_to_remove:
-                p.unlink(missing_ok=True)
-            _log_debug("task_delete: metadata files removed")
-        except Exception as exc:
-            _log_debug(f"task_delete: metadata removal failed: {exc}")
-            warnings.append(f"Metadata removal failed: {exc}")
-
-    if containers_removed:
-        _log_debug("task_delete: releasing web port")
-        try:
-            from ..ports import release_web_port
-
-            release_web_port(project.id, task_id)
-        except Exception as exc:  # noqa: BLE001 — best-effort cleanup
-            _log_debug(f"task_delete: web port release failed: {exc}")
-    else:
-        warnings.append("Web port kept claimed — containers may still be running")
-
-    _log_debug("task_delete: finished")
-    return TaskDeleteResult(task_id=task_id, warnings=warnings)
-
-
-def task_delete(project_id: str, task_id: str) -> TaskDeleteResult:
-    """Delete a task's workspace, metadata, and any associated containers.
-
-    Before removal, captures container logs and archives the task metadata
-    and logs to ``archive/<project_id>/tasks/``.  Containers are stopped
-    best-effort via podman using the ``<project.id>-<mode>-<task_id>``
-    naming scheme.  Returns a [`TaskDeleteResult`][terok.lib.orchestration.tasks.TaskDeleteResult] so the caller can
-    present any warnings from cleanup steps that failed.
-    """
-    return _task_delete(load_project(project_id), task_id)
-
-
-def _validate_login(project: ProjectConfig, task_id: str) -> tuple[str, str]:
-    """Validate that a task exists and its container is running.
-
-    Returns ``(container_name, mode)`` on success.
-    Raises ``SystemExit`` with actionable messages on failure.
-    """
-    meta_dir = tasks_meta_dir(project.id)
-    meta = read_task_meta(meta_dir, task_id)
-    if meta is None:
-        raise SystemExit(f"Unknown task {task_id}")
-
-    mode = meta.get("mode")
-    if not mode:
-        raise SystemExit(
-            f"Task {task_id} has never been run (no mode set).\n"
-            f"  Start a fresh task: terok task run {project.id}\n"
-            f"  Or run this stub:   terokctl task attach {project.id} {task_id} --mode cli"
-        )
-
-    cname = container_name(project.id, mode, task_id)
-    state = _rt.get_runtime().container(cname).state
-    if state is None:
-        raise SystemExit(
-            f"Container {cname} does not exist. "
-            f"Run 'terok task restart {project.id} {task_id}' first."
-        )
-    if state != "running":
-        raise SystemExit(
-            f"Container {cname} is not running (state: {state}). "
-            f"Run 'terok task restart {project.id} {task_id}' first."
-        )
-    return cname, mode
-
-
-def _get_login_command(project: ProjectConfig, task_id: str) -> list[str]:
-    """Return the command to interactively log into a task container."""
-    cname, _mode = _validate_login(project, task_id)
-    return _rt.get_runtime().container(cname).login_command()
-
-
-def _task_login(project: ProjectConfig, task_id: str) -> None:
-    """Open an interactive shell in a running task container."""
-    cmd = _get_login_command(project, task_id)
-    try:
-        os.execvp(cmd[0], cmd)
-    except FileNotFoundError:
-        raise SystemExit(
-            f"'{cmd[0]}' not found on PATH. Please install podman or add it to your PATH."
-        )
-
-
-def _task_stop(project: ProjectConfig, task_id: str, *, timeout: int | None = None) -> None:
-    """Gracefully stop a running task container."""
-    effective_timeout = timeout if timeout is not None else project.shutdown_timeout
-    meta_dir = tasks_meta_dir(project.id)
-    meta = read_task_meta(meta_dir, task_id)
-    if meta is None:
-        raise SystemExit(f"Unknown task {task_id}")
-    meta_file = meta_path(meta_dir, task_id)
-
-    mode = meta.get("mode")
-    if not mode:
-        raise SystemExit(f"Task {task_id} has never been run (no mode set)")
-
-    cname = container_name(project.id, mode, task_id)
-
-    state = _rt.get_runtime().container(cname).state
-    if state is None:
-        raise SystemExit(f"Task {task_id} container does not exist")
-    if state not in ("running", "paused"):
-        raise SystemExit(f"Task {task_id} container is not stoppable (state: {state})")
-
-    try:
-        _rt.get_runtime().container(cname).stop(timeout=effective_timeout)
-    except FileNotFoundError:
-        raise SystemExit("podman not found; please install podman")
-    except RuntimeError as exc:
-        raise SystemExit(f"Failed to stop container: {exc}")
-
-    try:
-        from ..ports import release_web_port
-
-        release_web_port(project.id, task_id)
-    except Exception:  # noqa: BLE001 — best-effort; container is already stopped
-        pass
-
-    from ..hooks import run_hook
-
-    run_hook(
-        "post_stop",
-        project.hook_post_stop,
-        project_id=project.id,
-        task_id=task_id,
-        mode=mode,
-        cname=cname,
-        task_dir=project.tasks_root / str(task_id),
-        meta_path=meta_file,
-    )
-
-    color_enabled = _supports_color()
-    print(f"Stopped task {task_id}: {_green(cname, color_enabled)}")
-    print(f"Restart with: terok task restart {project.id} {task_id}")
-
-
-def get_login_command(project_id: str, task_id: str) -> list[str]:
-    """Return the podman exec command to log into a task container."""
-    return _get_login_command(load_project(project_id), task_id)
-
-
-def task_login(project_id: str, task_id: str) -> None:
-    """Open an interactive shell in a running task container."""
-    _task_login(load_project(project_id), task_id)
-
-
-def task_stop(project_id: str, task_id: str, *, timeout: int | None = None) -> None:
-    """Gracefully stop a running task container.
-
-    Uses ``podman stop --time <N>`` to give the container *timeout* seconds
-    before SIGKILL.  When *timeout* is ``None`` the project's
-    ``run.shutdown_timeout`` setting is used (default 10 s).
-    """
-    _task_stop(load_project(project_id), task_id, timeout=timeout)
-
-
-def task_status(project_id: str, task_id: str) -> None:
-    """Show live task status with container state diagnostics."""
-    project = load_project(project_id)
-    meta_dir = tasks_meta_dir(project.id)
-    meta = read_task_meta(meta_dir, task_id)
-    if meta is None:
-        raise SystemExit(f"Unknown task {task_id}")
-
-    mode = meta.get("mode")
-    web_port = meta.get("web_port")
-    exit_code = meta.get("exit_code")
-
-    color_enabled = _supports_color()
-
-    # Query live container state
-    cname = None
-    cs = None
-    if mode:
-        cname = container_name(project.id, mode, task_id)
-        cs = _rt.get_runtime().container(cname).state
-
-    # Build TaskMeta for effective_status / mode_emoji computation
-    task = TaskMeta(
-        task_id=task_id,
-        mode=mode,
-        workspace=meta.get("workspace", ""),
-        web_port=web_port,
-        web_token=meta.get("web_token"),
-        backend=meta.get("backend"),
-        exit_code=exit_code,
-        deleting=bool(meta.get("deleting")),
-        initialized=_is_initialized(meta),
-        container_state=cs,
-        name=meta["name"],
-        provider=meta.get("provider"),
-        unrestricted=meta.get("unrestricted"),
-        created_at=meta.get("created_at"),
-    )
-    status = effective_status(task)
-    info = STATUS_DISPLAY.get(status, STATUS_DISPLAY["created"])
-
-    status_color = {"green": _green, "yellow": _yellow, "red": _red}.get(info.color, _yellow)
-    m = mode_info(task.mode)
-    m_emoji = render_emoji(m)
-
-    print(f"Task {task_id}:")
-    print(f"  Name:            {task.name}")
-    print(f"  Status:          {render_emoji(info)} {status_color(info.label, color_enabled)}")
-    print(f"  Mode:            {m_emoji} {m.label or 'not set'}")
-    if cname:
-        print(f"  Container:       {cname}")
-    if cs:
-        state_color = _green if cs == "running" else _yellow
-        print(f"  Container state: {state_color(cs, color_enabled)}")
-    elif mode:
-        print(f"  Container state: {_red('not found', color_enabled)}")
-    if task.unrestricted is not None:
-        perm_label = "unrestricted" if task.unrestricted else "restricted"
-        print(f"  Permissions:     {perm_label}")
-    if exit_code is not None:
-        print(f"  Exit code:       {exit_code}")
-    if web_port:
-        print(f"  Web port:        {web_port}")
-    # Work status from agent
-    tasks_root = project.tasks_root
-    agent_cfg = tasks_root / task_id / "agent-config"
-    ws = read_work_status(agent_cfg)
-    if ws.status:
-        print(f"  Work status:     {ws.status}")
-        if ws.message:
-            print(f"  Work message:    {ws.message}")
-
-
-# ---------- Archive operations ----------
-
-
-@dataclass
-class ArchivedTask:
-    """Metadata snapshot of an archived (deleted) task."""
-
-    archive_dir: Path
-    archived_at: str
-    task_id: str
-    name: str
-    mode: str | None
-    exit_code: int | None
-
-
-def _load_archived_task_meta(entry: Path) -> dict | None:
-    """Load an archived task's snapshot — JSON or legacy YAML — or ``None`` on miss.
-
-    Older archives wrote ``task.yml``; new ones write ``task.json``.  Both
-    are tolerated indefinitely on the read path because archives are
-    immutable once written and there's no operator action that would
-    rewrite them.
-    """
-    json_path = entry / "task.json"
-    if json_path.is_file():
-        try:
-            text = json_path.read_text(encoding="utf-8")
-            return json.loads(text) if text.strip() else {}
-        except (OSError, ValueError):
-            return None
-    yml_path = entry / "task.yml"
-    if not yml_path.is_file():
-        return None
-    try:
-        return _to_plain(_yaml_load(yml_path.read_text(encoding="utf-8")) or {})
-    except Exception:
-        return None
-
-
-def list_archived_tasks(project_id: str) -> list[ArchivedTask]:
-    """Return archived tasks for *project_id*, sorted newest-first."""
-    archive_root = tasks_archive_dir(project_id)
-    if not archive_root.is_dir():
-        return []
-    results: list[ArchivedTask] = []
-    for entry in sorted(archive_root.iterdir(), reverse=True):
-        if not entry.is_dir():
-            continue
-        meta = _load_archived_task_meta(entry)
-        if meta is None:
-            continue
-        # Parse archive timestamp from directory name: <timestamp>_<task_id>[_<name>]
-        parts = entry.name.split("_", 2)
-        archived_at = parts[0] if parts else entry.name
-        results.append(
-            ArchivedTask(
-                archive_dir=entry,
-                archived_at=archived_at,
-                task_id=str(meta.get("task_id", "")),
-                name=meta.get("name", ""),
-                mode=meta.get("mode"),
-                exit_code=meta.get("exit_code"),
-            )
-        )
-    return results
-
-
-def task_archive_list(project_id: str) -> None:
-    """Print archived tasks for *project_id*."""
-    archived = list_archived_tasks(project_id)
-    if not archived:
-        print("No archived tasks found")
-        return
-    for a in archived:
-        extra = []
-        if a.mode:
-            extra.append(f"mode={a.mode}")
-        if a.exit_code is not None:
-            extra.append(f"exit={a.exit_code}")
-        extra_s = f" [{'; '.join(extra)}]" if extra else ""
-        print(f"- {a.archived_at} #{a.task_id}: {a.name}{extra_s}")
-
-
-def task_archive_logs(project_id: str, archive_id: str) -> Path | None:
-    """Return the log file path for an archived task identified by *archive_id*.
-
-    *archive_id* is matched against archive directory names (prefix match).
-    Returns the log file path if found, or ``None``.
-    """
-    archive_root = tasks_archive_dir(project_id)
-    if not archive_root.is_dir():
-        return None
-    for entry in sorted(archive_root.iterdir(), reverse=True):
-        if entry.is_dir() and entry.name.startswith(archive_id):
-            log_file = entry / "logs" / "container.log"
-            if log_file.is_file():
-                return log_file
-    return None
+from .lifecycle import (  # noqa: F401 — re-exported public API
+    TaskDeleteResult,
+    capture_task_logs,
+    get_login_command,
+    task_delete,
+    task_login,
+    task_new,
+    task_rename,
+    task_status,
+    task_stop,
+)
+from .meta import (  # noqa: F401 — re-exported public API
+    CONTAINER_TEROK_CONFIG,
+    agent_config_dir,
+    dossier_path,
+    iter_task_ids,
+    load_task_meta,
+    mark_task_deleting,
+    meta_path,
+    read_task_meta,
+    task_exists,
+    tasks_archive_dir,
+    tasks_meta_dir,
+    update_task_exit_code,
+    write_task_meta,
+)
+from .naming import (  # noqa: F401 — re-exported public API
+    TASK_NAME_MAX_LEN,
+    generate_task_name,
+    sanitize_task_name,
+    validate_task_name,
+)
+from .query import (  # noqa: F401 — re-exported public API
+    TaskMeta,
+    get_all_task_states,
+    get_task_container_state,
+    get_task_meta,
+    get_tasks,
+    get_workspace_git_diff,
+    task_list,
+)
+
+__all__ = [
+    "ArchivedTask",
+    "CONTAINER_MODES",
+    "CONTAINER_TEROK_CONFIG",
+    "TASK_NAME_MAX_LEN",
+    "TaskDeleteResult",
+    "TaskMeta",
+    "agent_config_dir",
+    "capture_task_logs",
+    "container_name",
+    "dossier_path",
+    "generate_task_name",
+    "get_all_task_states",
+    "get_login_command",
+    "get_task_container_state",
+    "get_task_meta",
+    "get_tasks",
+    "get_workspace_git_diff",
+    "is_task_id",
+    "iter_task_ids",
+    "list_archived_tasks",
+    "load_task_meta",
+    "mark_task_deleting",
+    "meta_path",
+    "normalize_task_id_input",
+    "read_task_meta",
+    "resolve_task_id",
+    "sanitize_task_name",
+    "task_archive_list",
+    "task_archive_logs",
+    "task_delete",
+    "task_exists",
+    "task_list",
+    "task_login",
+    "task_new",
+    "task_rename",
+    "task_status",
+    "task_stop",
+    "tasks_archive_dir",
+    "tasks_meta_dir",
+    "update_task_exit_code",
+    "validate_task_name",
+    "write_task_meta",
+]
